requirements.txt
Flask==2.3.3
requests==2.31.0
python-telegram-bot==13.15  # Стабильная версия
gunicorn==21.2.0
python-dotenv==1.0.0
